% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_data.R
\name{annotate_data}
\alias{annotate_data}
\title{Annotate gene data}
\usage{
annotate_data(
  data,
  id.type = c("ensgene", "symbol", "entrez"),
  species = c("human", "mouse")
)
}
\arguments{
\item{data}{matrix or data.frame object whereby rownames are gene IDs}

\item{id.type}{gene ID type used for rownames.  May be one of \code{ensgene},
\code{symbol}, or \code{entrez}.}

\item{species}{name of the species. Only \code{'human'}, \code{'mouse'},
and \code{'rat'} are allowed.}
}
\description{
\code{annotate_data} annotates Ensembl, Entrez or Gene Symbol IDs.
}
