% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_index.R
\name{build_index}
\alias{build_index}
\title{Build index for mapping using Salmon}
\usage{
build_index(species = c("human", "mouse"), kmer = 31,
  ens_release = 93, dest.dir)
}
\arguments{
\item{species}{character string specifying the name of the species. Only
\code{'human'}, and \code{'mouse'} are supported at present.  [DEFAULT = human].}

\item{kmer}{k-mer size for indexing. default is 31. See \code{'Salmon'}
for details.}

\item{ens_release}{version of Ensembl release.}

\item{dest.dir}{directory where all the files will be saved.}
}
\value{
directory of index files
}
\description{
\code{build_index} for mapping reads using Salmon.
}
\examples{

#Running this function will take some time.
\donttest{
build_index(species="human",kmer=31,
ens_release=92, dest.dir=tempdir())
}

}
\references{
Rob Patro, Geet Duggal, Michael I. Love, Rafael A. Irizarry, and
Carl Kingsford (2017): Salmon provides fast and bias-aware
quantification of transcript expression. Nature methods, 14(4),417.
\url{https://www.nature.com/articles/nmeth.4197}
}
