% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sample.R
\name{check_sample}
\alias{check_sample}
\title{Check that targets file contains the necessary info in the right format}
\usage{
check_sample(sample.info, reference = NULL)
}
\arguments{
\item{sample.info}{character string giving the path to a tab-delimited text
file with at least the columns <condition> (treatment condition),
<sample> (sample name), and <file1> (absolute or relative path to the
fastq or salmon quant.sf files).  If fastq files and PE reads, then a column
<file2> should also be present.  If a batch effect is to be included in the
design, then this should be identified under the column <batch>.}

\item{reference}{character vector specifying the conditions in order.  For example,
c("A", "B", "C", "D") would mean "A" is the reference condition to which "B", "C"
and "D" are compared; in addition, "C" and "D" will be compared to "B", and "D"
will be compared to "C". If \code{NULL} then the comparisons will be arranged
alphabetically.  [DEFAULT = NULL].}
}
\description{
Check that targets file contains the necessary info in the right format
}
