% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_analysis.R
\name{edger_analysis}
\alias{edger_analysis}
\title{Run edgeR analysis on RNA-seq count data}
\usage{
edger_analysis(metadata, species = c("human", "mouse"))
}
\arguments{
\item{metadata}{list created by \code{sanity_check} and with raw
count data stored in \code{metadata$count$A} and length-adjusted
count data stored in \code{metadata$count$B}.  Created as part of
\code{run_dea}}

\item{species}{character string specifying the name of the species. Only
\code{'human'}, and \code{'mouse'} are supported at present.  [DEFAULT = human].}
}
\description{
Run edgeR analysis on RNA-seq count data
}
