% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis_deseq2.R
\name{exploratory_analysis_deseq2}
\alias{exploratory_analysis_deseq2}
\title{Exploratory analysis of DESeq2 data}
\usage{
exploratory_analysis_deseq2(dds, species, metadata)
}
\arguments{
\item{dds}{DESeqDataSet object created by \code{DESeq2::DESeq}}

\item{species}{character string specifying the name of the species. Only
\code{'human'}, and \code{'mouse'} are supported at present.  [DEFAULT = human].}

\item{metadata}{list created by \code{sanity_check} and with raw
count data stored in \code{metadata$count$A} and length-adjusted
count data stored in \code{metadata$count$B}.  Created as part of
\code{run_dea}}
}
\description{
Exploratory analysis of DESeq2 data
}
