% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis_edger.R
\name{exploratory_analysis_edger}
\alias{exploratory_analysis_edger}
\title{Exploratory analysis of edgeR data}
\usage{
exploratory_analysis_edger(disp, species, metadata)
}
\arguments{
\item{disp}{DGEList object; output of \code{edgeR::estimateDisp}}

\item{species}{character string specifying the name of the species. Only
\code{'human'}, and \code{'mouse'} are supported at present.  [DEFAULT = human].}

\item{metadata}{list created by \code{sanity_check} and with raw
count data stored in \code{metadata$count$A} and length-adjusted
count data stored in \code{metadata$count$B}.  Created as part of
\code{run_dea}}
}
\description{
Exploratory analysis of edgeR data
}
