% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_voom_analysis.R
\name{limma_voom_analysis}
\alias{limma_voom_analysis}
\title{Run Limma-voom analysis on RNA-seq count data}
\usage{
limma_voom_analysis(metadata, species = c("human", "mouse"))
}
\arguments{
\item{metadata}{list created by \code{sanity_check} and with raw
count data stored in \code{metadata$count$A} and length-adjusted
count data stored in \code{metadata$count$B}.  Created as part of
\code{run_dea}}

\item{species}{character string specifying the name of the species. Only
\code{'human'}, and \code{'mouse'} are supported at present.  [DEFAULT = human].}
}
\description{
Run Limma-voom analysis on RNA-seq count data
}
