% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Create heatmap with clustering of samples and genes for DEGs}
\usage{
make_heatmap(dds, DEG)
}
\arguments{
\item{dds}{DESeqDataSet object created by \code{DESeq2::DESeq}}

\item{DEG}{list of differential genes produced by \code{deseq2_analysis}}
}
\description{
Create heatmap with clustering of samples and genes for DEGs
}
