% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_PCA.R
\name{make_PCA}
\alias{make_PCA}
\title{Create PCA plot for transformed data}
\usage{
make_PCA(normcounts, proc, sample.info)
}
\arguments{
\item{normcounts}{normalised count data; created as part of
\code{edger_analysis}, \code{limma_voom_analysis}, or \code{deseq2_analysis}.}

\item{proc}{character string indicating analysis procedure used.  One of
"Limma-voom", "edgeR", "DESeq2". [DEFAULT = "Limma-voom].}

\item{sample.info}{metadata created by \code{sanity_check}}
}
\description{
Create PCA plot for transformed data
}
