% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scree.R
\name{make_scree}
\alias{make_scree}
\title{Create scree plot from prcomp data}
\usage{
make_scree(pca, proc)
}
\arguments{
\item{pca}{\code{prcomp} object.}

\item{proc}{character string indicating analysis procedure used.  One of
"Limma-voom", "edgeR", "DESeq2". [DEFAULT = "Limma-voom].}
}
\description{
Create scree plot from prcomp data
}
