% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_volcano.R
\name{make_volcano}
\alias{make_volcano}
\title{Create Volcano Plot}
\usage{
make_volcano(input, proc = c("Limma-voom", "edgeR", "DESeq2"))
}
\arguments{
\item{input}{result of differential analysis created as part of
\code{edger_analysis}, \code{limma_voom_analysis}, or \code{deseq2_analysis}}

\item{proc}{character string indicating analysis procedure used.  One of
"Limma-voom", "edgeR", "DESeq2". [DEFAULT = "Limma-voom].}
}
\description{
Create Volcano Plot
}
