% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_salmon.R
\name{read_salmon}
\alias{read_salmon}
\title{Read the run info from Salmon output}
\usage{
read_salmon(data.dir = NULL)
}
\arguments{
\item{data.dir}{Path to directories containing the Salmon results}
}
\description{
Function to import metadata from Salmon into R
}
\author{
Anil Chalisey
}
