% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fastqc.R
\name{run_fastqc}
\alias{run_fastqc}
\title{QC reports for fastq files using FastQC}
\usage{
run_fastqc(fastqc.files = NULL, dest.dir = "FASTQC", threads = NULL,
  fastqc = "fastqc")
}
\arguments{
\item{fastqc.files}{character vector of paths to fastq files. [DEFAULT = NULL].}

\item{dest.dir}{character string indicating directory to save results. [DEFAULT = "FASTQC"]}

\item{threads}{an integer value indicating the number of parallel threads to
be used by FastQC. [DEFAULT = maximum number of available threads - 1].}

\item{fastqc}{a character string specifying the path to the fastqc executable.
[DEFAULT = "fastqc"].}
}
\value{
\code{character} string specifying the path of the output directory.
}
\description{
Run the FastQC tool to generate HTML reports. FastQC may be installed from
\url{https://www.bioinformatics.babraham.ac.uk/projects/fastqc/}
}
\section{FastQC path}{

If the executable is in \code{$PATH}, then the default value for paths
(\code{"fastqc"}) will work. If it is not in \code{$PATH}, then the absolute
path should be given. If using Windows 10, it is assumed that FASTQC has
been installed in WSL, and the same rules apply.
}

\examples{
\dontrun{
fqFiles <- list.files(path = "fastqDir", pattern = "*.fastq", full.names = TRUE)
run_fastqc(fqFiles)
}

}
