% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multiqc.R
\name{run_multiqc}
\alias{run_multiqc}
\title{Generate MultiQC report for FastQC}
\usage{
run_multiqc(fastqc.dir, dest.dir, multiqc = "multiqc")
}
\arguments{
\item{fastqc.dir}{directory where all the FastQC files are saved.}

\item{dest.dir}{directory to save the combined QC report.}

\item{multiqc}{a character string specifying the path to the multiqc executable.
[DEFAULT = "multiqc"].}
}
\value{
HTML report.
}
\description{
\code{run_multiqc} generates a single HTML report from the FastQC results using MultiQC.
}
\section{MultiQC path}{

If the executable is in \code{$PATH}, then the default value for paths
(\code{"multiqc"}) will work. If it is not in \code{$PATH}, then the absolute
path should be given. If using Windows 10, it is assumed that MultiQC has
been installed in WSL, and the same rules apply.
}

\examples{

\dontrun{
run_multiqc(fastqc.dir=tempdir(), dest.dir=tempdir())
}

}
\references{
Philip Ewels, Mans Magnusson, Sverker Lundin, and Max Kaller (2016):
MultiQC: summarize analysis results for multiple tools and samples
in a single report. Bioinformatics, 32(19), 3047-3048.
\url{https://doi.org/10.1093/bioinformatics/btw354}
}
