% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tximport.R
\name{run_tximport}
\alias{run_tximport}
\title{Wrapper function to run tximport}
\usage{
run_tximport(
  species = c("human", "mouse"),
  salmon.files,
  summarise = TRUE,
  countsFromAbundance = c("no", "scaledTPM", "lengthScaledTPM")
)
}
\arguments{
\item{species}{character string specifying the name of the species. Only
\code{'human'}, and \code{'mouse'} are supported at present.  [DEFAULT = human].}

\item{salmon.files}{character vector specifying path to quant.sf files.}

\item{summarise}{logical, if \code{TRUE} then return gene-level summary, otherwise
return transcript level data.}

\item{countsFromAbundance}{whether to generate counts based on abundance.
Available options are: \code{'no'},
\code{'scaledTPM'} (abundance based estimated counts scaled up to
library size),
\code{'lengthScaledTPM'}.}
}
\value{
a list of gene and transcript level estimated counts.
}
\description{
\code{run_tximport} function runs tximport on transcript level
abundances from Salmon to summarise to gene level. See Bioconductor
package \link[tximport]{tximport} for details.
}
\details{
We use Ensembl annotation for both genes and transcripts.
}
\references{
Charlotte Soneson, Michael I. Love, Mark D. Robinson (2015):
Differential analyses for RNA-seq: transcript-level estimates
improve gene-level inferences. F1000Research.
\url{http://dx.doi.org/10.12688/f1000research.7563.1}
}
