% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_check.R
\name{sanity_check}
\alias{sanity_check}
\title{Import the metadata and check it is correct}
\usage{
sanity_check(
  sample.info,
  reference = NULL,
  species = c("human", "mouse"),
  output.dir,
  threads = NULL
)
}
\arguments{
\item{sample.info}{character string giving the path to a tab-delimited text
file with at least the columns <condition> (treatment condition),
<sample> (sample name), and <file1> (absolute or relative path to the
fastq or salmon quant.sf files).  If fastq files and PE reads, then a column
<file2> should also be present.  If a batch effect is to be included in the
design, then this should be identified under the column <batch>.}

\item{reference}{character vector specifying the conditions in order.  For example,
c("A", "B", "C", "D") would mean "A" is the reference condition to which "B", "C"
and "D" are compared; in addition, "C" and "D" will be compared to "B", and "D"
will be compared to "C". If \code{NULL} then the comparisons will be arranged
alphabetically.  [DEFAULT = NULL].}

\item{species}{character string specifying the name of the species. Only
\code{'human'}, and \code{'mouse'} are supported at present.  [DEFAULT = human].}

\item{output.dir}{character string specifying the directory to which results
will be saved.  If the directory does not exist, it will be created.}

\item{threads}{an integer value indicating the number of parallel threads to
be used by FastQC. [DEFAULT = maximum number of available threads - 1].}
}
\value{
A \code{data.frame}
}
\description{
Import the metadata and check it is correct
}
