% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.R
\name{pathway_analysis}
\alias{pathway_analysis}
\title{Title}
\usage{
pathway_analysis(x, species = c("human", "mouse"))
}
\arguments{
\item{x}{output from \code{edger_analysis()}, \code{deseq2_analysis},
\code{limma_voom_analysis}.}

\item{species}{character string specifying the name of the species. Only
\code{'human'}, and \code{'mouse'} are supported at present.  [DEFAULT = human].}
}
\description{
Title
}
